backend/requirements.txt
django==3.2.12
djangorestframework==3.12.0
django-cors-headers==3.11.0
PyJWT==2.4.0
django-tenant-schemas==1.11.0
pillow==9.4.0
django-admin-volt==1.0.4
PyQRCode==1.2.1
django-ckeditor==6.5.1
fcm-django==1.0.12
firebase-admin==5.4.0
weasyprint==54.0
whitenoise==6.5.0